[refactor] ChessPiece 의 중복되는 코드 Override 로 변경
